source/notebooks/sagemaker_predictive_maintenance_entry_point/requirements.txt
mxnet-cu90mkl>=1.4
gluonnlp
pandas==0.22